+++ src/main.tsx
@@ -4,15 +4,18 @@ import App from "./App";
 import { CrawlEventsProvider } from "./hooks/useCrawlEvents";
 import { ToastProvider } from "./hooks/useToasts";
 import { ThemeProvider } from "./hooks/useTheme";
+import { EscapeStackProvider } from "./contexts/EscapeStack";
 import "./styles/index.css";
 
 ReactDOM.createRoot(document.getElementById("root")!).render(
   <React.StrictMode>
     <ToastProvider>
       <ThemeProvider>
-        <CrawlEventsProvider>
-          <App />
-        </CrawlEventsProvider>
+        <EscapeStackProvider>
+          <CrawlEventsProvider>
+            <App />
+          </CrawlEventsProvider>
+        </EscapeStackProvider>
       </ThemeProvider>
     </ToastProvider>
   </React.StrictMode>